app_telegrambot/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohttp==3.10.10
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
annotated-types==0.7.0
certifi==2024.8.30
frozenlist==1.4.1
multidict==6.1.0
magic-filter==1.0.12
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
requests==2.32.3